requirements.txt
annotated-types==0.7.0
anthropic==0.94.0
anyio==4.12.1
async-timeout==4.0.3
certifi==2026.2.25
charset-normalizer==3.4.7
distro==1.9.0
docstring_parser==0.17.0
exceptiongroup==1.3.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.14.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.28
langchain-anthropic==0.3.22
langchain-core==0.3.84
langchain-text-splitters==0.3.11
langsmith==0.4.37
orjson==3.11.5
packaging==25.0
pydantic==2.13.0
pydantic_core==2.46.0
python-dotenv==1.2.1
PyYAML==6.0.3
requests==2.32.5
requests-toolbelt==1.0.0
sniffio==1.3.1
SQLAlchemy==2.0.49
tenacity==9.1.2
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.1
zstandard==0.25.0
